Some general Scryfall UI tweaks
