+++ src/scryfall/scryfallCardFound.ts
@@ -1,127 +1,6 @@
-import { Message, AttachmentBuilder, EmbedBuilder } from "discord.js";
-import { Card, Cards, Prices } from "scryfall-api";
-import { PricingData } from "../types/scryfall/PricingData";
-import { convertPricesToGBP, getLowestHighestData } from "./scryfallHelpers";
-import { getImageUrl } from "./scryfallImageHelpers";
+import { Message } from "discord.js";
 import { isSendableChannel } from "../util/typeGuards";
-import moment from "moment-timezone";
-
-export function pricesToGBP(prices: Prices): string {
-  if (!(prices.usd || prices.usd_foil || prices.eur || prices.eur_foil)) {
-    return "???";
-  }
-
-  const conversionRate: number = prices.usd || prices.usd_foil ? 0.75 : 0.86;
-  const price: string =
-    prices.usd ?? prices.usd_foil ?? prices.eur ?? prices.eur_foil ?? "0";
-
-  const GBPPrice = parseFloat(price!) * conversionRate;
-  return GBPPrice > 100 ? GBPPrice.toString() : to2DP(GBPPrice);
-}
-
-export function to2DP(number: number): string {
-  return (Math.round(number * 100) / 100).toFixed(2);
-}
-
-export async function getCardMessageObject(
-  message: Message,
-  cardName: string,
-  set: string | undefined = undefined,
-  number: number | undefined = undefined
-): Promise<any> {
-  if (!isSendableChannel(message.channel)) return;
-
-  const cardDetails: Card | undefined = number
-    ? await Cards.bySet(set!, number)
-    : await Cards.byName(cardName, set, false);
-
-  if (!cardDetails) {
-    await message.channel.send(
-      `Ran into an error fetching ${cardName} for set ${set} and number ${number}!`
-    );
-    return;
-  }
-
-  const [isImageLocal, imageUrl]: [boolean, string] = await getImageUrl(
-    cardDetails
-  );
-  const attachment: AttachmentBuilder | null = isImageLocal
-    ? new AttachmentBuilder(`${imageUrl}.jpg`)
-    : null;
-  const foilOnly: boolean =
-    cardDetails.prices.usd === null && cardDetails.prices.usd_foil !== null;
-  const releaseDate = moment(cardDetails.released_at);
-  const unreleased: boolean = releaseDate.isAfter(moment.now());
-
-  const legality: string =
-    cardDetails.legalities.commander === "legal"
-      ? "Legal"
-      : `Non-legal${unreleased ? "*" : ""}`;
-  const formattedPrice: number = Math.min(
-    ...convertPricesToGBP(cardDetails.prices)
-  );
-  const price: string = `£${
-    formattedPrice !== Infinity ? to2DP(formattedPrice) : "???"
-  }${foilOnly ? " (F)" : ""}`;
-  const rarity: string =
-    cardDetails.rarity.charAt(0).toUpperCase() + cardDetails.rarity.slice(1);
-  const spacer: number = Math.floor(
-    (32 - legality.length - price.length - rarity.length) / 2
-  );
-
-  const footer: string = `${legality}${" ".repeat(
-    spacer + 1
-  )}${price}${" ".repeat(Number(unreleased) + spacer - 1)}${rarity}\n${
-    cardDetails.set_name
-  } (${cardDetails.set})${
-    unreleased ? `\n\n*Releases on ${releaseDate.format("Do MMM YYYY")}` : ""
-  }`;
-
-  const embed: EmbedBuilder = new EmbedBuilder()
-    .setTitle(cardDetails.name)
-    .setURL(cardDetails.scryfall_uri)
-    .setFooter({ text: footer });
-
-  if (imageUrl) {
-    // EmbedBuilder.setImage actually checks that this is a valid URI!!!
-    embed.setImage(
-      isImageLocal ? `attachment://${imageUrl.split("/").pop()}.jpg` : imageUrl
-    );
-  }
-
-  const oracleId: string =
-    cardDetails.oracle_id ?? cardDetails.card_faces?.[0].oracle_id ?? "";
-  if (oracleId.length) {
-    const lowestHighestData: PricingData | undefined =
-      await getLowestHighestData(oracleId);
-
-    if (lowestHighestData) {
-      embed.addFields({
-        name: "Prices",
-        value:
-          lowestHighestData.lowestPrice !== Infinity &&
-          lowestHighestData.highestPrice !== -Infinity
-            ? `
-Lowest: [${lowestHighestData.lowestSet} @ \
-£${to2DP(lowestHighestData.lowestPrice)}](${lowestHighestData.lowestUrl})
-Highest: [${lowestHighestData.highestSet} @ \
-£${to2DP(lowestHighestData.highestPrice)}](${lowestHighestData.highestUrl})
-`
-            : "No pricing data found!",
-      });
-    }
-  } else {
-    console.error(
-      `Couldn't find an Oracle ID for card named ${cardDetails.name}, ID ${cardDetails.id}!`
-    );
-  }
-
-  return {
-    content: cardDetails.scryfall_uri.replace("?utm_source=api", ""),
-    embeds: [embed],
-    files: attachment ? [attachment] : [],
-  };
-}
+import { getCardMessageObject } from "./scryfallEmbedObjectBuilder";
 
 export async function scryfallCardFound(
   message: Message,
