+++ src/scryfall/scryfallNoCardFound.ts
@@ -1,12 +1,12 @@
 import { Message, EmbedBuilder } from "discord.js";
 import { isSendableChannel } from "../util/typeGuards";
 
-export function scryfallNoCardFound(message: Message, cardName: string): void {
+export async function scryfallNoCardFound(message: Message, cardName: string): Promise<void> {
   if (!isSendableChannel(message.channel)) return;
 
   const embed: EmbedBuilder = new EmbedBuilder().setDescription(
     `No card found for "${cardName}"`
   );
 
-  message.channel.send({ embeds: [embed] });
+  await message.channel.send({ embeds: [embed] });
 }
